Move syncer to sync package
