+++ pkg/registry/apis/provisioning/register.go
@@ -31,6 +31,7 @@ import (
 	listers "github.com/grafana/grafana/pkg/generated/listers/provisioning/v0alpha1"
 	"github.com/grafana/grafana/pkg/registry/apis/provisioning/auth"
 	"github.com/grafana/grafana/pkg/registry/apis/provisioning/jobs"
+	"github.com/grafana/grafana/pkg/registry/apis/provisioning/jobs/sync"
 	"github.com/grafana/grafana/pkg/registry/apis/provisioning/repository"
 	"github.com/grafana/grafana/pkg/registry/apis/provisioning/repository/github"
 	"github.com/grafana/grafana/pkg/registry/apis/provisioning/resources"
@@ -432,6 +433,7 @@ func (b *APIBuilder) GetPostStartHooks() (map[string]genericapiserver.PostStartH
 			}
 			b.repositoryLister = repoInformer.Lister()
 
+			syncer := sync.NewSyncer(b.parsers, b.resourceLister)
 			b.jobs.Register(jobs.NewJobWorker(
 				b,
 				b.parsers,
@@ -439,6 +441,7 @@ func (b *APIBuilder) GetPostStartHooks() (map[string]genericapiserver.PostStartH
 				b.identities,
 				b.render,
 				b.resourceLister,
+				syncer,
 				b.blobstore,
 				b.urlProvider,
 			))
